+++ trunk/src/faust-cc/crybaby.cc
@@ -3,20 +3,20 @@ namespace crybaby {
 
 FAUSTFLOAT 	fslider0;
 FAUSTFLOAT 	fslider1;
-float 	fConst0;
-float 	fConst1;
-float 	fRec1[2];
-float 	fRec2[2];
-float 	fRec3[2];
+double 	fConst0;
+double 	fConst1;
+double 	fRec1[2];
+double 	fRec2[2];
+double 	fRec3[2];
 FAUSTFLOAT 	fslider2;
-float 	fRec0[3];
+double 	fRec0[3];
 int	fSamplingFreq;
 
 void init(int samplingFreq)
 {
 	fSamplingFreq = samplingFreq;
-	fConst0 = (1413.7166941154069f / fSamplingFreq);
-	fConst1 = (2827.4333882308138f / fSamplingFreq);
+	fConst0 = (1413.7166941154069 / fSamplingFreq);
+	fConst1 = (2827.4333882308138 / fSamplingFreq);
 	for (int i=0; i<2; i++) fRec1[i] = 0;
 	for (int i=0; i<2; i++) fRec2[i] = 0;
 	for (int i=0; i<2; i++) fRec3[i] = 0;
@@ -25,23 +25,23 @@ void init(int samplingFreq)
 
 void compute(int count, float *input0, float *output0)
 {
-	float 	fSlow0 = fslider0;
-	float 	fSlow1 = (1 - max(0, fSlow0));
-	float 	fSlow2 = fslider1;
-	float 	fSlow3 = powf(2.0f,(2.3f * fSlow2));
-	float 	fSlow4 = (1 - (fConst0 * (fSlow3 / powf(2.0f,(1.0f + (2.0f * (1.0f - fSlow2)))))));
-	float 	fSlow5 = (0.0010000000000000009f * (0 - (2.0f * (cosf((fConst1 * fSlow3)) * fSlow4))));
-	float 	fSlow6 = (0.0010000000000000009f * faustpower<2>(fSlow4));
-	float 	fSlow7 = (0.0001000000000000001f * powf(4.0f,fSlow2));
-	float 	fSlow8 = fslider2;
-	float 	fSlow9 = (1 - max(0, (0 - fSlow0)));
+	double 	fSlow0 = fslider0;
+	double 	fSlow1 = (1 - max(0, fSlow0));
+	double 	fSlow2 = fslider1;
+	double 	fSlow3 = pow(2.0,(2.3 * fSlow2));
+	double 	fSlow4 = (1 - (fConst0 * (fSlow3 / pow(2.0,(1.0 + (2.0 * (1.0 - fSlow2)))))));
+	double 	fSlow5 = (0.0010000000000000009 * (0 - (2.0 * (cos((fConst1 * fSlow3)) * fSlow4))));
+	double 	fSlow6 = (0.0010000000000000009 * faustpower<2>(fSlow4));
+	double 	fSlow7 = (0.0001000000000000001 * pow(4.0,fSlow2));
+	double 	fSlow8 = fslider2;
+	double 	fSlow9 = (1 - max(0, (0 - fSlow0)));
 	for (int i=0; i<count; i++) {
-		float fTemp0 = (float)input0[i];
-		fRec1[0] = (fSlow5 + (0.999f * fRec1[1]));
-		fRec2[0] = (fSlow6 + (0.999f * fRec2[1]));
-		fRec3[0] = (fSlow7 + (0.999f * fRec3[1]));
-		fRec0[0] = ((fSlow8 * (fTemp0 * fRec3[0])) - (0.996f * ((fRec2[0] * fRec0[2]) + (fRec1[0] * fRec0[1]))));
-		output0[i] = (FAUSTFLOAT)((fSlow9 * (fRec0[0] - (0.996f * fRec0[1]))) + (fSlow1 * fTemp0));
+		double fTemp0 = (double)input0[i];
+		fRec1[0] = (fSlow5 + (0.999 * fRec1[1]));
+		fRec2[0] = (fSlow6 + (0.999 * fRec2[1]));
+		fRec3[0] = (fSlow7 + (0.999 * fRec3[1]));
+		fRec0[0] = ((fSlow8 * (fTemp0 * fRec3[0])) - (0.996 * ((fRec2[0] * fRec0[2]) + (fRec1[0] * fRec0[1]))));
+		output0[i] = (FAUSTFLOAT)((fSlow9 * (fRec0[0] - (0.996 * fRec0[1]))) + (fSlow1 * fTemp0));
 		// post processing
 		fRec0[2] = fRec0[1]; fRec0[1] = fRec0[0];
 		fRec3[1] = fRec3[0];
@@ -53,9 +53,9 @@ void compute(int count, float *input0, float *output0)
 static struct RegisterParams { RegisterParams(); } RegisterParams;
 RegisterParams::RegisterParams()
 {
-	registerVar("crybaby.level","","S","",&fslider2, 0.1f, 0.0f, 1.0f, 0.01f);
-	registerVar("crybaby.wah","","S","",&fslider1, 0.0f, 0.0f, 1.0f, 0.01f);
-	registerVar("crybaby.wet_dry","wet/dry","S","",&fslider0, 0.0f, -1.0f, 1.0f, 0.1f);
+	registerVar("crybaby.level","","S","",&fslider2, 0.1, 0.0, 1.0, 0.01);
+	registerVar("crybaby.wah","","S","",&fslider1, 0.0, 0.0, 1.0, 0.01);
+	registerVar("crybaby.wet_dry","wet/dry","S","",&fslider0, 0.0, -1.0, 1.0, 0.1);
 	registerInit(init);
 }
