+++ trunk/src/faust-cc/feed.cc
@@ -3,9 +3,9 @@ namespace feed {
 
 FAUSTFLOAT 	fslider0;
 FAUSTFLOAT 	fslider1;
-float 	fVec0[4];
+double 	fVec0[4];
 FAUSTFLOAT 	fslider2;
-float 	fRec0[6];
+double 	fRec0[6];
 int	fSamplingFreq;
 
 void init(int samplingFreq)
@@ -17,12 +17,12 @@ void init(int samplingFreq)
 
 void compute(int count, float *input0, float *output0, float *output1)
 {
-	float 	fSlow0 = min(0.999999f, fslider0);
-	float 	fSlow1 = fslider1;
-	float 	fSlow2 = fslider2;
-	float 	fSlow3 = ngate;
+	double 	fSlow0 = min(0.999999, fslider0);
+	double 	fSlow1 = fslider1;
+	double 	fSlow2 = fslider2;
+	double 	fSlow3 = ngate;
 	for (int i=0; i<count; i++) {
-		float fTemp0 = (float)input0[i];
+		double fTemp0 = (double)input0[i];
 		fVec0[0] = fTemp0;
 		fRec0[0] = fold(fSlow0, (fSlow3 * ((fVec0[0] + (fSlow2 * fVec0[3])) - (fSlow1 * fRec0[5]))));
 		output0[i] = (FAUSTFLOAT)fRec0[0];
@@ -36,9 +36,9 @@ void compute(int count, float *input0, float *output0, float *output1)
 static struct RegisterParams { RegisterParams(); } RegisterParams;
 RegisterParams::RegisterParams()
 {
-	registerVar("amp.feedforward","Feedforward","S","",&fslider2, 0.0f, -1.0f, 1.0f, 0.01f);
-	registerVar("amp.feedback","Feedback","S","",&fslider1, 0.0f, -1.0f, 1.0f, 0.01f);
-	registerVar("amp.fuzz","fuzz","S","",&fslider0, 1.0f, 0.0f, 1.0f, 0.01f);
+	registerVar("amp.feedforward","Feedforward","S","",&fslider2, 0.0, -1.0, 1.0, 0.01);
+	registerVar("amp.feedback","Feedback","S","",&fslider1, 0.0, -1.0, 1.0, 0.01);
+	registerVar("amp.fuzz","fuzz","S","",&fslider0, 1.0, 0.0, 1.0, 0.01);
 	registerInit(init);
 }
