+++ trunk/src/faust-cc/inputgain.cc
@@ -2,7 +2,7 @@ namespace inputgain {
 // generated from file '../src/faust/inputgain.dsp'
 
 FAUSTFLOAT 	fslider0;
-float 	fRec0[2];
+double 	fRec0[2];
 int	fSamplingFreq;
 
 void init(int samplingFreq)
@@ -13,10 +13,10 @@ void init(int samplingFreq)
 
 void compute(int count, float *input0, float *output0)
 {
-	float 	fSlow0 = (0.0010000000000000009f * powf(10,(0.05f * fslider0)));
+	double 	fSlow0 = (0.0010000000000000009 * pow(10,(0.05 * fslider0)));
 	for (int i=0; i<count; i++) {
-		fRec0[0] = (fSlow0 + (0.999f * fRec0[1]));
-		output0[i] = (FAUSTFLOAT)((float)input0[i] * fRec0[0]);
+		fRec0[0] = (fSlow0 + (0.999 * fRec0[1]));
+		output0[i] = (FAUSTFLOAT)((double)input0[i] * fRec0[0]);
 		// post processing
 		fRec0[1] = fRec0[0];
 	}
@@ -25,7 +25,7 @@ void compute(int count, float *input0, float *output0)
 static struct RegisterParams { RegisterParams(); } RegisterParams;
 RegisterParams::RegisterParams()
 {
-	registerVar("amp.in_level","in / level","S","",&fslider0, 0.0f, -4e+01f, 4e+01f, 0.1f);
+	registerVar("amp.in_level","in / level","S","",&fslider0, 0.0, -4e+01, 4e+01, 0.1);
 	registerInit(init);
 }
